producers/requirements.txt
feedparser
kafka-python